Fix: split oversized health, queue, and workflow test files (#84)
Reduce maintenance risk by extracting the monolithic health route, download queue service internals, and full workflow Playwright suite into focused modules while preserving behavior and public contracts.

Changes:
- Split health route types, probes, and status aggregation into dedicated modules
- Split queue persistence and scheduler logic from DownloadQueueService API surface
- Split full-workflow Playwright coverage into focused spec files for submission, queue updates, cancel/error, and accessibility/integration

Fixes #84
